fix(types):add default export
